Make ParseSess an optional arg to FormatHandler::handle_formatted_file
The only implementaiton of the `FormatHandler` trait is the one
implemented for `Session`, and in that implementation we pass the
`ParseSess` to `source_file::write_file`, which also takes the
`ParseSess` as an optional argument.

looking at the implementation of `source_file::write_file`, the
`ParseSess` is optionally used to determine the file's original line
ending instead of reading the file from the file system again to do so.

Given the above explanation, I don't think there's any issue changing
the signature of `FormatHandler::handle_formatted_file` and the
implementation of `FormatHandler for Session`.